Renamed the project from *SignPy* to *Chirper*
