+++ setup.py
@@ -4,14 +4,14 @@ with open("README.md", "r") as file:
     long_description = file.read()
 
 setuptools.setup(
-    name="signpy2",
+    name="chirper-py",
     version="1.0",
     author="Cristobal Allendes",
     author_email="callendes.molina@gmail.com",
     description="Package for signal analysis",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/No-tengo-nombre/signpy",
+    url="https://github.com/No-tengo-nombre/chirper",
     classifiers=[
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",
