+++ src/chirper/modulation/fm.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from signpy.sgn import Signal1
-from signpy.config import HERTZ, FM_MODULATION
+from chirper.sgn import Signal1
+from chirper.config import HERTZ, FM_MODULATION
 
 
 def fm_modulation(signal1: Signal1, *args, method=FM_MODULATION, hertz=HERTZ, **kwargs) -> Signal1:
