+++ src/chirper/modulation/pm.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from signpy.sgn import Signal1
-from signpy.config import PM_MODULATION, HERTZ
+from chirper.sgn import Signal1
+from chirper.config import PM_MODULATION, HERTZ
 
 
 def pm_modulation(signal1: Signal1, *args, method=PM_MODULATION, hertz=HERTZ, **kwargs) -> Signal1:
